close #4
This commit finalizes the main signal handler. There is too much state
to manage trying to handle a HUP ourselves, so now whenever we receive
any of the following signals:

  - SIGUP
  - SIGINT
  - SIGTERM
  - SIGQUIT

we just kill all child processes, write a few log messages, and exit.
When running as a daemon, we can just SIGTERM and rerun the binary
to restart.